+++ golang_gin_sql/.env
@@ -1,5 +1,5 @@
 PORT=3001
-MODE=debug
+GIN_MODE=debug
 DB_USER=user
 DB_PASSWORD=password
 DB_NAME=companies
