--- golang_gin_sql/cmd/server/start
@@ -1,3 +1,0 @@
-#!/bin/bash
-
-gin --build . --path ../..
