+++ golang_gin_sql/docker-compose.ci.yml
@@ -9,7 +9,7 @@ services:
       - 8080:8080
     environment:
       PORT: 8080
-      MODE: release
+      GIN_MODE: release
       DB_USER: user
       DB_PASSWORD: password
       DB_NAME: companies
