+++ golang_gin_sql/run
@@ -1,0 +1,3 @@
+#!/bin/bash
+
+gin --build ./cmd/server --path .
